requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.10.2
    # via -r requirements.in
anyio==3.6.2
    # via starlette
asyncpg==0.27.0
    # via -r requirements.in
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via uvicorn
cryptography==40.0.1
    # via pyjwt
dnspython==2.3.0
    # via email-validator
email-validator==1.3.1
    # via fastapi-users
fastapi==0.95.0
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-users[sqlalchemy]==10.4.1
    # via
    #   -r requirements.in
    #   fastapi-users-db-sqlalchemy
fastapi-users-db-sqlalchemy==5.0.0
    # via fastapi-users
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   requests
jinja2==3.1.2
    # via -r requirements.in
loguru==0.6.0
    # via -r requirements.in
makefun==1.15.1
    # via fastapi-users
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
passlib[bcrypt]==1.7.4
    # via
    #   -r requirements.in
    #   fastapi-users
psycopg2-binary==2.9.6
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via fastapi
pyhumps==3.8.0
    # via -r requirements.in
pyjwt[crypto]==2.6.0
    # via
    #   -r requirements.in
    #   fastapi-users
python-multipart==0.0.6
    # via
    #   -r requirements.in
    #   fastapi-users
requests==2.28.2
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlalchemy[asyncio]==2.0.8
    # via
    #   alembic
    #   fastapi-users-db-sqlalchemy
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via
    #   alembic
    #   pydantic
    #   sqlalchemy
urllib3==1.26.15
    # via requests
uvicorn==0.21.1
    # via -r requirements.in
